trustledger-backend/requirements.txt
pathway>=0.7.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.0
openai>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
